requirements.txt
aiogram==3.0.0b7
telethon==1.40.0
redis
python-dotenv>=1.0.0
cryptg
Pillow
pycryptodome
openpyxl==3.1.2
pytest>=7.0.0